refactor: prepare for color generator
With this commit, the color generator /should/ be able to generate both
the new and the old color variables, thus allowing for a much simpler
and cleaner implementation since it will be using only what it needs.

I'm only doing /primary/ colors in this update because I have yet to
figure out what to do with /secondary/ colors.